+++ include/aws/crt/mqtt/Mqtt5Client.h
@@ -474,7 +474,7 @@ namespace Aws
                  *
                  * @return this option object
                  */
-                Mqtt5ClientOptions &WithPort(uint16_t port) noexcept;
+                Mqtt5ClientOptions &WithPort(uint32_t port) noexcept;
 
                 /**
                  * Set booststrap for mqtt5 client
